feat: merge final a main de la rama feature/view

+++ src/main/java/model/Empleado.java
@@ -17,7 +17,7 @@ public class Empleado extends Persona implements Serializable {
     public String getCodigoEmpleado() { return codigoEmpleado; }
 
 
-    public Empleado(String nombre, String nombreUsuario, String contraseña, String documentoid, String email, String rol, int telefono, String codigoEmpleado, String cargo, double salario) {
+    public Empleado(String nombre, String nombreUsuario, String contraseña, String documentoid, String email, String rol, long telefono, String codigoEmpleado, String cargo, double salario) {
         super(nombre, nombreUsuario, contraseña, documentoid, email, rol, telefono);
         this.codigoEmpleado = codigoEmpleado;
         this.cargo = cargo;
